Updated openstack/openstack Project: openstack/barbican  4e9cbe27ce9b22a88ed14da19ceee2cd2d639c96
Fixing unauthenticated middleware role bug

We regressed a few days to where the unauthenticated middleware
context is completely broken. This patch attempts to resolve the
regression through two ways. The first is to provides a way to specify
the roles through the X-Roles header. It is assumed that a deployer
using their own middleware should be providing this header. The second
is that if the X-Roles header isn't specified then the admin role is
assumed; normally this is when using barbican in dev mode.

Change-Id: Ic19270f0381b4aec2e0c8a4ce466900bc91e1597
Closes-Bug: 1455247