Updated openstack/openstack Project: openstack-infra/zuul  5a412bfc497998c4eef381bbf292fc5784dfee90
Simplify cloner required parameters

Prior to this change you not only had to always supply zuul-* args or
set ZUUL_* env vars but you had to also set zuul params that were
unused. Trim down the required zuul params to only those that are used
(branch, url, ref) and only require all of them if one is set.

zuul-cloner is a useful tool for configuring git repos from a git cache
even without zuul params, this change exposes that functionality. In
particular this is handy if we ever want to run zuul-cloner outside of a
CI context, for example on a developers laptop.

Change-Id: I4f285b489df64007a6ec4c14522623f38939c74b