Updated openstack/openstack Project: openstack/neutron-fwaas  f5f28ec6fa46fcb4009c2105bf09dfdbe0b7041b
Remove contextlib.nested from tests

With the deprecation of this method in Python2.7, the neutron hacking
check has been modified to check for this (N324). Since this check has
been upstreamed (I171fbdb89892a3d4548bf2ca52f4a7dd9ef8dccb), and the
neutron-fwaas repo uses the same hacking check, the FWaaS test cases
are now failing.
This commit does the same changes as was done in neutron to remove use
of this method.

Partial-Bug: #1428424

Change-Id: I130cf3eaeccbf6a965407c76a74f8f8fa59943a4