Updated openstack/openstack Project: openstack-infra/shade  6d654bc7adce911d34216b2038ef0e02cc6dba1c
Catch client exceptions during list ops

We were not consistently catching the underlying client exceptions
from the list operations, and just letting them escape as non-wrapped
exceptions. This corrects that.

Also, list_images() was trying to catch an OpenStackCloudException
that was never going to happen from the code in the try block.

Change-Id: Ia863ec2f988c7e90138097be15cd66fc858d1b21