Updated openstack/openstack Project: openstack/nova  9f581dd81918a1c6cb674b77f28691783641d807
Add migration_type and hidden to Migration database model

This adds an enumeration for migration_type to the schema. It does not
do any data translation in the schema migration, and allows the column
to be nullable so that we don't have to act on existing migrations.
This means that any outstanding migrations in the database will need to
be handled lazily by the upper layers.

This also adds a flag called hidden, defaulting to False. Since we are
using this new type field to record potentially admin-initiated actions,
it is desirable to allow them to be hidden from regular users at times.
This adds that field to the database so that we can support that in the
migration APIs.

Related to blueprint robustify-evacuate

Change-Id: I133d46d146415e19f5c120dfa3e37090291b1298