Updated openstack/openstack Project: openstack/heat  54f1ad927e684a993e247311f564897a1b63ed62
Pass stack to thread in resource_signal

This is the only thread that gets started without an explicit reference to
the stack. This prevents us from using a weakref in the resource, as it
will cause the Stack's reference count to hit 0 before the thread is run.
Now we explicitly pass a reference to the stack.

Change-Id: Ie51be7b54d97ef184e401e395a7e7e3a26ce003b
Related-Bug: #1454873