Updated openstack/openstack Project: openstack/barbican  32380f1fd7fa8acd949b861e3f49cb88cf9989d9
Fix the sql_pool_logging config attribute

The 'sql_pool_logging' config field is a boolean value, but it was
being treated as a text value in repositories.py, so any value
supplied to it was treated as a non-None condition which enabled
SQLAlchemy pool logging always. This CR changes this field to the
proper boolean type.

Change-Id: If2c1471d5a3b1fca4ac5033754d05b386fb0cbf5
Closes-Bug: #1455493