Updated openstack/openstack Project: openstack/nova  98e44ebac79c609956ce3a85d7c13b9ad2440bd5
Add migration_type to Migration object

From the records that could be in the database prior to now, we can
determine their type by the flavor ids on the record. So, we do that
live-ly so that we don't need to (a) bother layers above with the potential
for a migration_type=None and (b) do the conversion in a db migration.

Note that this includes a change to the API to not show this field to
the user. It will be exposed in v2.1 in a later patch.

Related to blueprint robustify-evacuate

Change-Id: Id78c91357c892e684cfd8eeb1f000a57180865bb