Updated openstack/openstack Project: openstack/nova  13d9a5bac58ddd87a763c13de616a5584bb53182
libvirt: deprecate libvirt version usage < 0.10.2

In the 2016.1 'M' release we're going to raise the required minimum
version of libvirt to 0.10.2 so add a check in init_host such that if
the version of libvirt being used is less than 0.10.2 we log a warning
that people need to move up when going to 2016.1.

According to the libvirt version distro support matrix:

https://wiki.openstack.org/wiki/LibvirtDistroSupportMatrix

This will drop support for Debian Wheezy (if not using
Wheezy-backports), openSUSE 12.2 and Ubuntu Precise 12.04 (if not using
the cloud archive).

UpgradeImpact: The minimum required version for libvirt in the 2016.1
               'M' release will be 0.10.2.

Change-Id: Ia257780355101b7970df4a79b9ea0007d1fbb041