Updated openstack/openstack Project: openstack/nova  97296ec125794124620bb1f7a7c23eef9f64e442
Fix race between resource audit and cpu pinning

This fixes a race between the claimed CPUs of an instance being
persisted to the DB, and the resource audit scanning the DB for
instances and subtracting pinned CPUs from the list of available CPUs.

The problem only shows up when the following sequence happens:
1) instance A (with dedicated cpus) boots on a compute node
2) resource audit runs on that compute node
3) instance B (with dedicated cpus) boots on the same compute node

The bug is that the claimed numa topology isn't updated until
after we release COMPUTE_RESOURCES_SEMAPHORE, so when the resource
audit retrieves the list of instances the numa_topology hasn't
been updated yet for the most recent one.

The fix is to persist the claimed numa topology before releasing
COMPUTE_RESOURCES_SEMAPHORE.

Closes-Bug: #1454451
Co-Authored-By: Dan Smith <dansmith@redhat.com>
Change-Id: I553f2e43a68577c83d890c3671380af68f9e725a