Updated openstack/openstack Project: openstack/nova  c13f693d496f4ea5dbc78244fefd9ec2d6ea8de5
Detach volume after deleting instance with no host

If an instance is booted from a volume, shelved, and goes into an error
state due to some reason. Volume from which instance is booted, remains
in-use state even the instance is deleted because instance has no host
associated with it.

Called _local_delete() to detach volume and destroy bdm if instance is
in shelved_offloaded state or has no host associated with it. This will
cleanup both volumes and the networks.

Currently in test_servers.py, "test_delete_server_instance" executes
similar to "test_delete_server_instance_while_building". This is because
"test_delete_server_instance" calls instance.save() method which updates
vm_state to building where it should be in active state.

Fixed "test_delete_server_instance" to test deleting an instance which
is in active state and has a valid host.

Closes-Bug: #1404867
Closes-Bug: #1408527
Change-Id: Ic630ae7d026a9697afec46ac9ea40aea0f5b5ffb