Updated openstack/openstack Project: openstack-dev/devstack  c9e842871b850da24a57e2ba5c98020c529111b2
Don't set tempest fixed_network_name with neutron

If neutron is enabled then there isn't a shared private network
between all tenants which is what is required for the
fixed_network_name config option. This commit adds a conditional
to not set that option when neutron is enabled. While not necessarily
fatal to tempest it does emit a warning on almost every server create
call if it is set with a non-existent network name.

Change-Id: I1a42fa6b0b5a93b411c08ec35df043d6ea69d453