Updated openstack/openstack Project: openstack/tripleo-incubator  4a6265e5d50c2efe59406f7af167d4282295f912
Add support for overriding cache just for seed

When testing with the seed as the undercloud, I'm frequently wanting
to rebuild just the seed VM image, e.g when I know the overcloud images
are good, but I want to test a new heat patch.

This adds support for SEED_USE_CACHE, which will override the global
USE_CACHE variable if set.

Change-Id: I81f256feedbee26a258333a68bf8715bbc7710bf