Updated openstack/openstack Project: openstack/nova  8e4c665e01a920f09beb03e56c12bf2eeff1fb65
objects: convert VirtCPUModel to use named enums

The VirtCPUModel class currently uses anonymous enums for a
number of fields. Since some of the same enums will be required
in other objects, convert them to named enums instead.

The import of nova.compute.arch and nova.compute.cpumodel is
a temporary state. The intention is that the constants in those
two modules actually get moved into the enum classes themselves
allowing the modules to be removed. This is to be done in a
later commit.

The hash of VirtCPUModel changes because the field classname
has changed. The version is not incremented though, because
this is an ABI-compatible change.

Change-Id: Id79793510e02d47f9ef377b7b3a9ffa08dff4a66