Updated openstack/openstack Project: openstack/tempest-lib  f93048334cf9f93a823dd42387f76ecddf27d30d
Pass response body down into RestClientExceptions

This commit adds support to RestClientException objects to store the
response body as a local attr. Additionally it modifies all the uses
of the subclassed RestClientExceptions in the base rest client to
store the response body on each error type.

This is done in order to check the content in some occasion that
additional error details are coming back from a service in the body.
It will be useful to have the resp_body passed into the exception so
that it can be used in additional checks.

Change-Id: Ie1bb259e7e683081f0ad127617acc8deb98467c0