Updated openstack/openstack Project: openstack-infra/elastic-recheck  993f0cb0c6107e571acfbe45d03ef46ffbe473ab
Add query for bug 1456771

Since May 15th 2015 I sometimes see failures in both
check and gate pipelines with "Failure prepping block
device in the gate".

2771 hits in 7 days, check and gate, not 100% failure
because this is a delete path which is not explicitly
tested for most scenarios, so it's only really a "fail"
in explicit delete server tests.

https://review.openstack.org/#/c/145738/ appears to
have introduced the race.

Related-Bug: #1456771

Change-Id: I1e9392067f19f1fbd86c5407debf936a9fdf8228