Updated openstack/openstack Project: openstack/ha-guide  e78694afbc0b5167bc00c44c3b90f3a997a91d48
HA Concepts for introduction

This commit mostly contains material that was in the old ha-guide
with a few edits.  It adds new material about quorums and
single-controller ha environments.

We should add links to other parts of the documentation as those
pieces are written.  We also have an outstanding LP for additional
information that is required.

Note that some of these terms are defined in the Glossary and the docs
now have links to those terms.  We may want to augment those Glossary
definitions as we move forward.

Implements blueprint improve-ha-guide

Change-Id: I01ca79b5694d5ffd4a81f1984d98d36238e5c880