Updated openstack/openstack Project: openstack/ironic  fe2f445418ad7cb7741a7d237fd2ad69c27833a4
Add simplegeneric to py34 requirements

simplegeneric suppoosed to be installed by WSME, but for a some reasons
it was missed in py3 env. We can't add this library to requirements.txt,
because it missed in global requirements, so I've added simplegeneric
to the `deps` section in py34 environment.

This if a fast-and-dirty way to unblock python34 gates for ironic.

Change-Id: Ib0cddf7541defbdfe1b17448b52becf8bd4ef3de