Updated openstack/openstack Project: openstack/oslo.db  9fe59b628b3fe77e571f1667e9477b37460217de
Remove pre-SQLAlchemy-0.9.7 compat utilities

Now that requirements.txt is at SQLAlchemy 0.9.7 at the minimum,
we can remove the engine_connect compatibility module as well
as part of the handle_error compatibility module.

In particular, this gives us the ability to add and remove
engine_connect events using the event API directly which
can be handy in test fixtures.

Change-Id: I48450ad9d472d4377913ad391a0f5e3ba0f1471f