Updated openstack/openstack Project: openstack/oslo-specs  ed569d61b45d9d6e953bafef95ea4b06515ee9e9
oslo.messaging: remove double reply

Currently, when we are  waiting for a RPC reply, for each call we
receive two AMQP messages - first one with the payload, a second one
to ensure the other have finish to send the payload.

We are going to remove the second message for RPC reply.

Co-Authored-By: Victor Sergeyev <vsergeyev@mirantis.com>

Change-Id: I42106f4948279984b4dc037ca17eb9bc1cbc8207