Updated openstack/openstack Project: openstack-infra/tripleo-ci  32f867e1e45d2c9f9e18362fd6a2eed2d16337f2
puppet: use a single image for CI

This patch updates the Puppet CI job so that it uses the new
disk_image.yaml functionality in devtest to use a single
image for all of the Overcloud roles. This should help
save CI time and also significantly cleans up how
we configure the puppet CI job.

Depends-On: I1b1dbd5a7e2a4fda2eff31813518df8bc72e7e56

Change-Id: I11c2674391834ec09e0b029a3f9b181b8456de8b