Updated openstack/openstack Project: openstack/nova  0edfed4187dc42d79f39825e9c6907ca24bafd1a
Revert "Detach volume after deleting instance with no host"

This reverts commit d1baa9fe7eb342b63fc85cbb5ef70bb676de6566

The change introduced a race on delete where there isn't a host while
prepping block devices and we hit a NoneType error in nova-compute.

There was also a recheck grind on the original change showing it wasn't
safe and needs to be reworked.

Change-Id: Id4e405e7579530ed1c1f22ccc972d45b6d185f41
Closes-Bug: #1456771
Related-Bug: #1448316