Updated openstack/openstack Project: openstack/tripleo-puppet-elements  c5c349d49b1e9a194026d046337a3ffa20f562b8
Rename $name variable to $module_name

dib-run-parts also uses a $name variable as part of its log output.
Since $name was being set here, it caused the name of the last listed
puppet module (puppet-tuskar in this case) to be logged on every line of
output from the entire diskimage-builder run. Use $module_name to avoid
the variable name collision.

Change-Id: Ibd57f757fbe5a97592869e8244b9fb4b1a310298