Updated openstack/openstack Project: openstack/python-keystoneclient  841e634aceae0b483fb32f3cd9b8e3a99d662bad
Support discovery on the AUTH_INTERFACE

We need to allow get_endpoint(interface=auth.AUTH_INTERFACE, version=X)
to support the same version negotiation that the service catalog goes
through. This is required to support generic plugins where you often
provide an unversioned auth_url to the plugin but need a versioned URL
to query for available projects.

Change-Id: Id423a538c169264a81c5714e6a9eff9b33912a55
Closes-Bug: #1438013