Updated openstack/openstack Project: openstack/neutron  d825ef947fc275da2f0b4b1f2fd2695347c492be
Catch ObjectDeletedError and skip port or subnet removal

When network is deleted service ports are deleted in the scope of
delete_network. Service ports could also be deleted by other entities
such as DHCP agent releasing dhcp port.
That could rarely lead to a race condition when port object used in
_delete_ports helper is already deleted causing ObjectDeletedError
exception.
Need to handle it and prevent object deletion in that case.

Change-Id: I531251d3211545c82a5bb7a471b7915da9b763b7
Closes-Bug: #1454408