Updated openstack/openstack Project: openstack/diskimage-builder  eb732d3d65fbb920be5068b06a97bceec0e27041
Initial element tests

Adding a test function which allows us to use elements to perform
element-specific tests. In order for this to work sanely, also adding
some configuration to our break system so we can assert on negative
tests.

Also adding a test for apt-sources to verify this code actually works.

Change-Id: I378a74255010eca192f5766b653f8a42404be5ea