Updated openstack/openstack Project: openstack/python-barbicanclient  5f1a164c8986ab925bf5fd41f23e95175d63a222
Remove instances of _base_url

The variable _base_url was not really necessary since we are setting up
the endpoint_override variable which will eventually be used when making
a request through the keystoneclient's session.

This is a step in getting the client to use the keystoneclient's
API more wisely. A subsequent CR will remove the need for the
endpoint_override and make use of keystoneclient's endpoint_filters to
discover the Barbican endpoint.

Change-Id: Id589c76fb5fe481a3dd48ee1baa26f7d19d2d1a1