Updated openstack/openstack Project: openstack/keystonemiddleware  e7f634e9d57c316eb6cd6c73c0427d8a37b2635c
Cleanup token hashes generated by cache

We hash a token under multiple configurable algorithms for revocations
and caching and so must check all of them. This hash generation was
being done and returned by the cache check and reused in revocation
check. This is an unusual pattern and requires the cache object to have
knowledge of token types and how to hash them.

Change validate so that we generate the hash values in the main function
and pass that to the cache and revocation functions. This moves the
function to get the first available token id from the list to the main
file. This means the cache interface is much more normal with get and
set id functions and encapsulates the hash list function in the one
file.

Change-Id: I9194dbd052674f64122ff74329ce292a342512d3