Updated openstack/openstack Project: openstack/python-openstackclient  7cc9632939ba05381a50faca581ffd2cb22351b3
Fix insecure/verify options

The insecure and verify options are broken so that verify always
gets set to True.  One problem was that the parsed args not
defaulted so os_cloud_config thinks there was always a command
line specified.  The other problem was getattr was called on cloud
config instead of get.

Closes-Bug: #1450855
Change-Id: Ib5f004f51a7453cc8f5a89759e2031ec42e04a30