Updated openstack/openstack Project: openstack/api-wg  dd86e696e49409dedbd516d6432cb39ed95c0b5c
Add the description of GET method

I guess everyone is thinking the same thing as this about GET method
and it would be unnecessary to describe it on the guideline.
However, there are some APIs which change resource state actually
on multiple(not single) projects. For example, Nova contains
* startup a host
* shutdown a host
* reboot a host
with GET method.

This is a small try before changing them.

Change-Id: Icd91ec3b3716f6d0c9f90a4a7ad4a4da41840168