Updated openstack/openstack Project: openstack/python-keystoneclient  4776495adfadbf5240a9e0f169990ce139af9549
Provide a means to get all installed plugins

Particular for use in writing error messages and help text it can be
useful to get a list of all the plugins that are installed on the
system. Provide a version that returns the classes as well so that you
don't have to reload the modules if the user is picking one.

Closes-Bug: #1423711
Change-Id: I021249eac8156c2d3ccbbacb7503184b6eb6e784