Updated openstack/openstack Project: openstack/oslo.utils  7d335e3810a7d10289b8ad08df4f3fbf8ff6938b
timeutils: utcnow() can return a value with a timezone

utcnow() used to return a non-timezone-aware timezone, which is terrible
as there is then no way to know if it's a local time object or a UTC
time zone object. This patch changes utcnow() so it can carry that
information in the datetime object.

Change-Id: Ib098ef16063d97e37ff2b9a0fc934f67fd9ed5f3