Updated openstack/openstack Project: openstack-infra/project-config  ef393bb145bfff4ec58285f71d98abf6df043254
Don't run kolla functional gate as root

Because of a recent change, docker image building cannot be done
as the root user.  This is a good thing ;)  But breaks the current
gate.

Also we have a setup_docker.sh script already in our tools directory,
so rename the script we are running to setup_gate.sh because that is
actually what the script does.

Change-Id: I8f86616fb19b4b03f34fe40acf1689f6b174ff98