Updated openstack/openstack Project: openstack/cinder  40dca69de02be495190b8aa7949e778e353f367c
Add ability for drivers to copy data preserving sparseness

Drivers can set self._sparse_copy_volume_data to opt in
to copying data in a way that preserves sparseness.

This is likely to be desired elsewhere, but currently this
patch only enables it for the NFS driver and provides a
method for other drivers to opt in, in case there are
unexpected issues in some drivers.

On the NFS driver, this prevents volume migration from
changing a volume file from sparse to fully allocated.

Change-Id: Ib74b2fb4015b805d33de42dfe60c8a97cc970b84
Closes-Bug: #1455595