Updated openstack/openstack Project: openstack/nova  00b5935f8973b2762f6185d5c5aa35304b5763d1
Run tests with PyMySQL on Python 3

Add test-requirements-py3.txt to use PyMySQL, instead of MySQL-python,
on Python 3.

Add a py34 test environment to tox.ini which uses
test-requirements-py3.txt and sets OS_TEST_DBAPI_ADMIN_CONNECTION to use
the PyMySQL connector instead of the MySQL-python connector.

This change is based on Ironic tox.ini, so I added Victor Sergeyev as
the coauthor of the patch.

Co-Authored-By: Victor Sergeyev <vsergeyev@mirantis.com>
Blueprint nova-python3
Change-Id: I5a27c7d627105b614da3cab3a87296a8e186b555